Deal with "the" in titles
On the HMH Book Expert Online site, they place "the" at the end of the
book title on the page (and therefore in the HTML). "The" in the English
language is something called an article which is a modifier that appears
before nouns or noun phrases.

For a concrete example, consider the book "The Abduction". Google Books
API would return the title as "The Abduction", the sanitize function
would lowercase everything and remove the spaces and then attempt to
look for a match to "theabduction" on the HMH site. It wouldn't find a
match because the data structure that is built from the HMH book titles
would contain "abductionthe" (the same sanitization gets applied to
remove spaces and lowercase).

Now both sanitization methods remove "the" from the beginning or end of
the book title to increase the chance of finding a match between the
scanned book title and the results on the HMH page.

This commit also adds some extra characters that get removed from titles
such as "-" and "'".